Updates to remove obstacle type in preference of 2D and 3D shape types
