+++ Anytime_RRT_Sharp/src/ArrtsService.hpp
@@ -6,7 +6,8 @@
 #include "ConfigspaceGraph.hpp"
 #include "ConfigspaceNode.hpp"
 #include "cppshrhelp.hpp"
-#include "Geometry.hpp"
+#include "Geometry2D.hpp"
+#include "Geometry3D.hpp"
 #include "WorkspaceGraph.hpp"
 
 using namespace std;
