+++ Anytime_RRT_Sharp/src/CMakeLists.txt
@@ -16,8 +16,8 @@ add_library(ConfigspaceGraph ConfigspaceGraph.cpp)
 add_library(ConfigspaceNode ConfigspaceNode.cpp)
 add_library(WorkspaceGraph WorkspaceGraph.cpp)
 add_library(Vehicle Vehicle.cpp)
-add_library(Obstacle Obstacle.cpp)
-add_library(Geometry Geometry.cpp)
+add_library(Geometry2D Geometry2D.cpp)
+add_library(Geometry3D Geometry3D.cpp)
 add_library(ArrtsEngine ArrtsEngine.cpp)
 add_library(ArrtsParams ArrtsParams.cpp)
 add_library(ArrtsService ArrtsService.cpp)
@@ -27,8 +27,8 @@ list(APPEND EXTRA_LIBS ConfigspaceGraph)
 list(APPEND EXTRA_LIBS ConfigspaceNode)
 list(APPEND EXTRA_LIBS WorkspaceGraph)
 list(APPEND EXTRA_LIBS Vehicle)
-list(APPEND EXTRA_LIBS Obstacle)
-list(APPEND EXTRA_LIBS Geometry)
+list(APPEND EXTRA_LIBS Geometry2D)
+list(APPEND EXTRA_LIBS Geometry3D)
 list(APPEND EXTRA_LIBS ArrtsEngine)
 list(APPEND EXTRA_LIBS ArrtsParams)
 list(APPEND EXTRA_LIBS ArrtsService)
@@ -38,8 +38,8 @@ list(APPEND TEST_LIBS ConfigspaceGraph)
 list(APPEND TEST_LIBS ConfigspaceNode)
 list(APPEND TEST_LIBS WorkspaceGraph)
 list(APPEND TEST_LIBS Vehicle)
-list(APPEND TEST_LIBS Obstacle)
-list(APPEND TEST_LIBS Geometry)
+list(APPEND TEST_LIBS Geometry2D)
+list(APPEND TEST_LIBS Geometry3D)
 list(APPEND TEST_LIBS ArrtsParams)
 list(APPEND TEST_LIBS gtest)
