+++ Anytime_RRT_Sharp/src/ConfigspaceGraph.hpp
@@ -4,7 +4,8 @@
 #include <vector>
 #include "cppshrhelp.hpp"
 #include "ConfigspaceNode.hpp"
-#include "Geometry.hpp"
+#include "Geometry2D.hpp"
+#include "Geometry3D.hpp"
 
 using namespace std;
