+++ Anytime_RRT_Sharp/src/Geometry2D.cpp
@@ -1,4 +1,4 @@
-#include "Geometry.hpp"
+#include "Geometry2D.hpp"
 
 Vector::Vector()
 {
@@ -119,7 +119,7 @@ Point Plane::point() const { return _point; }
 
 Vector Plane::normal() const { return _normal; }
 
-Point Plane::getIntersectionPoint(Line line) const
+Point Plane::getIntersectionPoint(Line& line) const
 {
     Vector u = line.p2() - line.p1();
     Vector w = line.p1() - point();
@@ -130,137 +130,13 @@ Point Plane::getIntersectionPoint(Line line) const
 
     if (denom == 0)
     {
-        Point *p;
-        p = NULL;
+        Point *p = NULL;
         return *p;
     }
-    else
-    {
-        double s = numer / denom;
-        return line.p1() + u * s;
-    }
+
+    double s = numer / denom;
+    return line.p1() + u * s;
 }
-
-Sphere::Sphere()
-{
-    _x = 0;
-    _y = 0;
-    _z = 0;
-    _radius = 0;
-    _area = 0;
-}
-
-Sphere::Sphere(Point p, double radius)
-{
-    _x = p.x();
-    _y = p.y();
-    _z = p.z();
-    _radius = radius;
-    _area = _calculateVolume();
-}
-
-Sphere::Sphere(double x, double y, double z, double radius)
-{
-    _x = x;
-    _y = y;
-    _z = z;
-    _radius = radius;
-    _area = _calculateVolume();
-}
-
-double Sphere::_calculateVolume() const
-{
-    return (4.0/3.0) * M_PI * pow(radius(), 3);
-}
-
-double Sphere::radius() const { return _radius; }
-
-double Sphere::volume() const { return _area; }
-
-Rectangle::Rectangle()
-{
-    _minPoint = Point();
-    _maxPoint = Point();
-    _volume = 0;
-}
-
-Rectangle::Rectangle(Point minPoint, Point maxPoint)
-{
-    _minPoint = minPoint;
-    _maxPoint = maxPoint;
-    _volume = _calculateVolume();
-    _surfaces = _calculateSurfaces();
-    _points = _calculatePoints();
-}
-
-Rectangle::Rectangle(double minX, double minY, double minZ, double maxX, double maxY, double maxZ)
-{
-    _minPoint = Point(minX, minY, minZ);
-    _maxPoint = Point(maxX, maxY, maxZ);
-    _volume = _calculateVolume();
-    _surfaces = _calculateSurfaces();
-    _points = _calculatePoints();
-}
-
-double Rectangle::_calculateVolume() const
-{
-    double xDiff = maxX() - minX();
-    double yDiff = maxY() - minY();
-    double zDiff = maxZ() - minZ();
-    return xDiff * yDiff * zDiff;
-}
-
-vector<Plane> Rectangle::_calculateSurfaces() const
-{
-    vector<Plane> surfaces(6);
-    surfaces[0] = Plane(Point(minX(), 0, 0), Vector(1, 0, 0));
-    surfaces[1] = Plane(Point(maxX(), 0, 0), Vector(1, 0, 0));
-    surfaces[2] = Plane(Point(0, minY(), 0), Vector(0, 1, 0));
-    surfaces[3] = Plane(Point(0, maxY(), 0), Vector(0, 1, 0));
-    surfaces[4] = Plane(Point(0, 0, minZ()), Vector(0, 0, 1));
-    surfaces[5] = Plane(Point(0, 0, maxZ()), Vector(0, 0, 1));
-    return surfaces;
-}
-
-vector<Point> Rectangle::_calculatePoints() const
-{
-    vector<Point> points(8);
-    points[0] = Point(minX(), minY(), minZ());
-    points[1] = Point(maxX(), minY(), minZ());
-    points[2] = Point(minX(), maxY(), minZ());
-    points[3] = Point(minX(), minY(), maxZ());
-    points[4] = Point(maxX(), maxY(), minZ());
-    points[5] = Point(minX(), maxY(), maxZ());
-    points[6] = Point(maxX(), minY(), maxZ());
-    points[7] = Point(maxX(), maxY(), maxZ());
-    return points;
-}
-
-Point Rectangle::minPoint() const { return _minPoint; }
-
-Point Rectangle::maxPoint() const { return _maxPoint; }
-
-const vector<Plane>& Rectangle::surfaces() const { return _surfaces; }
-
-const Plane& Rectangle::surfaces(int i) const { return _surfaces[i]; }
-
-const vector<Point>& Rectangle::points() const { return _points; }
-
-const Point& Rectangle::points(int i) const { return _points[i]; }
-
-double Rectangle::volume() const { return _volume; }
-
-double Rectangle::minX() const { return _minPoint.x(); }
-
-double Rectangle::minY() const { return _minPoint.y(); }
-
-double Rectangle::minZ() const { return _minPoint.z(); }
-
-double Rectangle::maxX() const { return _maxPoint.x(); }
-
-double Rectangle::maxY() const { return _maxPoint.y(); }
-
-double Rectangle::maxZ() const { return _maxPoint.z(); }
 
 State::State()
 {
