+++ Anytime_RRT_Sharp/src/WorkspaceGraph.hpp
@@ -1,6 +1,6 @@
 #include <vector>
-#include "Geometry.hpp"
-#include "Obstacle.hpp"
+#include "Geometry2D.hpp"
+#include "Geometry3D.hpp"
 #include "Vehicle.hpp"
 
 #ifndef WORKSPACE_H
@@ -9,12 +9,12 @@
 class WorkspaceGraph : public Rectangle
 {
     GoalState _goalRegion;
-    vector<SphereObstacle> _obstacles;
+    vector<Sphere> _obstacles;
     Vehicle _vehicle;
     void _buildWorkspaceGraph();
     bool _goalRegionReached;
     bool _obstacleInFreespace(double x, double y, double z, double radius) const;
-    bool _obstacleInFreespace(SphereObstacle o) const;
+    bool _obstacleInFreespace(Sphere o) const;
 
     public:
 
@@ -23,15 +23,15 @@ class WorkspaceGraph : public Rectangle
         GoalState goalRegion();
         void setGoalRegion(double x, double y, double z, double theta, double radius);
         void setGoalRegion(State goalState, double radius);
-        vector<SphereObstacle> obstacles();
-        SphereObstacle obstacles(int i);
+        vector<Sphere> obstacles();
+        Sphere obstacles(int i);
         void defineFreespace(Rectangle limits);
         bool checkAtGoal(GraphNode node);
         bool nodeIsSafe(Point p);
         bool pathIsSafe(Point p1, Point p2);
         bool pathIsSafe(Point p1, vector<Point> points);
         void addObstacle(double x, double y, double z, double radius);
-        void addObstacles(vector<SphereObstacle> obstacles);
+        void addObstacles(vector<Sphere> obstacles);
         bool atGate(GraphNode node);
         WorkspaceGraph() { _buildWorkspaceGraph(); }
 };
